feat(member): enhance office presence functionality with session management and toggle state handling
